rabbitmq/requirements.txt
pika==0.10.0
mysql-connector==2.1.4
